Allow multi user
Hello administrator, it would be great if yo could apply this simple change that we are using daily. 
If the .lock file is stored in /tmp it doesn’t allow several users to use kubelogin at the same time in the same computer, by storing the .lock file in the home directory of each user, we allow multiple users to use kubelogin on the same computer.
This change has been tested by me in a production environment.
Have a good day. Juan.