Add --browser-args option
This change adds the --browser-args option, which takes a slice of strings.
These arguments will be passed to the browser command specified with
--browser-command. This allows for prepending arguments before the URL.

The primary use case for this feature is to allow opening a browser without
needing to make an alias or script to add arguments.
For example, "powershell.exe start $URL" allows opening a browser in WSL
without needing to access the registry to read the default browser.